requirements.txt
asgiref==3.10.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.0
Django==6.0.1
django-pwa==2.0.1
djangorestframework==3.16.1
idna==3.11
pillow==12.1.0
python-decouple==3.8
reportlab==4.4.6
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.3
urllib3==2.6.2
webencodings==0.5.1
whitenoise==6.11.0
